escha W2: both gates PASS; the bit-exact gate caught a bug in the reference
  decode gate: 24/24 tiles BIT-EXACT   (both bit rates, four projection types, real tiles)
  gemm gate:   10/10 cases at rel 1.6-1.7e-3, M in {5,8,40,64}, KS in {1,4}

1.6e-3 is the bf16 output rounding floor (2^-9) -- the same floor the MXFP4 and int4 kernels sit
at -- so the GEMM is exact to the representable precision, and DKS=1 and DKS=4 agree.

THE REFERENCE WAS WRONG AND THE DEVICE WAS RIGHT. First run mismatched ~53% of values, but case 1
showed got=2.90234 against want=2.90137: half an fp16 ulp, not a wrong value. The codebook's final
step is `lo + hi` and CUDA does it with __hadd, in fp16; the NumPy reference widened both halves to
float32 first, so its sum was too accurate. Fixed to add in fp16.

Worth recording for two reasons. It is invisible to every test used to recover the format --
correlation, std, sorted values are all unmoved by half an ulp, so the earlier +0.907/+0.976
validation stands unchanged -- and it would have been invisible here too under any tolerance. Only
demanding bit-exactness of the DECODE surfaced it. That is exactly why the decode gate is bit-exact
while the GEMM gate is not: the GEMM legitimately needs a tolerance (fp32 WMMA accumulate against
an fp64 reference), the decode legitimately does not.

+++ escha/exl3_ref.py
@@ -48,9 +48,12 @@ def decode_3inst(state, cb=1):
     else:
         raise ValueError(cb)
     x = ((x & np.uint32(0x8FFF8FFF)) ^ np.uint32(0x3B603B60)).astype(np.uint32)
-    lo = (x & 0xFFFF).astype(np.uint16).view(np.float16).astype(np.float32)
-    hi = (x >> 16).astype(np.uint16).view(np.float16).astype(np.float32)
-    return lo + hi
+    lo = (x & 0xFFFF).astype(np.uint16).view(np.float16)
+    hi = (x >> 16).astype(np.uint16).view(np.float16)
+    # The add is fp16, NOT fp32. CUDA does __hadd here, so widening first would make the reference
+    # disagree with every real decoder by up to half an fp16 ulp -- which the bit-exact device gate
+    # duly caught (got 2.90234 vs want 2.90137). Harmless for correlation, fatal for a bit gate.
+    return (lo + hi).astype(np.float32)
 
 
 def tile_states(words, K):
